Moved functions back to ImGuilmpl.cpp
