--- soh/soh/Enhancements/applypresets.h
@@ -1,8 +1,0 @@
-#pragma once
-
-void applyEnhancementPresets(void);
-void applyEnhancementPresetDefault(void);
-void applyEnhancementPresetVanillaPlus(void);
-void applyEnhancementPresetEnhanced(void);
-void applyEnhancementPresetRandomizer(void);
-void applyEnhancementPresetIronman(void);
